Splitting constants.py
